feat(storage): SQLite → PostgreSQL migration support
Translator (adapters/storage/pg_adapter.py):
- BIGINT widening for telegram_id/chat_id/sent_to/message_id/*_by columns
  in schema.py / platform_schema.py / tenant_schema.py to fix int4 overflow
  on Telegram IDs > 2^31.
- _PgConnection: asyncio.Lock around every self._conn.* call. asyncpg
  connections aren't safe for concurrent use; without the lock parallel
  API requests collide with "another operation is in progress".
- executescript: strip -- comments before splitting on `;` so a stray
  semicolon inside a comment doesn't tear a CREATE TABLE in two.
- New translations: PRAGMA table_info(X) → information_schema query
  with the same 6-column shape; sqlite_master → pg_tables/pg_indexes
  UNION exposing (name, type, sql).

Migrations:
- migrate_email_unique_per_account: fixed whitespace-mismatch in the
  UNIQUE(account_id,email) skip-check that caused repeated destructive
  rebuilds (whitespace-stripped both sides now).
- New migrate_widen_telegram_id_columns: ALTER COLUMN TYPE BIGINT for
  prod cutover; no-op on SQLite, idempotent on PG.
- platform_geofences in legacy migrations.py uses BIGINT for created_by.

Core (core.py):
- _initialize_pg also runs platform_schema/migrations so PG gets the
  full table set (subscriptions, billing_usage_snapshots, ai_chat_history,
  error_log, role_ai_guidance) that legacy SQLite had via prior runs.
- Fixed 60s pool.close() hang: release writer connection before
  pool.close() so asyncpg doesn't wait command_timeout for in-flight drain.

Migration tooling:
- scripts/export_sqlite_to_postgres.py: drop legacy SQLite-only columns
  (resolved_at, completion_notes, dismissed, last_odometer) at insert
  time so a stale SQLite column doesn't fail every row.
- New scripts/validate_pg_parity.py: row-count diff between SQLite and
  PG with per-table OK/DRIFT/MISSING/EXTRA verdicts.
- New tests/test_pg_adapter_translator.py: 27 tests pinning every
  translation rule (placeholders, AUTOINCREMENT→SERIAL, datetime/date
  rewrites, INSERT OR IGNORE, real-codebase SQL samples).

+++ adapters/storage/platform_schema.py
@@ -31,7 +31,7 @@ async def create_tables(conn) -> None:
 
         CREATE TABLE IF NOT EXISTS users (
             id              INTEGER PRIMARY KEY AUTOINCREMENT,
-            telegram_id     INTEGER NOT NULL UNIQUE,
+            telegram_id     BIGINT  NOT NULL UNIQUE,
             account_id      INTEGER NOT NULL REFERENCES accounts(id),
             role            TEXT    NOT NULL DEFAULT 'fleet',
             department      TEXT    NOT NULL DEFAULT 'general',
@@ -79,7 +79,7 @@ async def create_tables(conn) -> None:
         CREATE TABLE IF NOT EXISTS authorized_chats (
             id          INTEGER PRIMARY KEY AUTOINCREMENT,
             account_id  INTEGER NOT NULL REFERENCES accounts(id),
-            chat_id     INTEGER NOT NULL,
+            chat_id     BIGINT  NOT NULL,
             chat_title  TEXT    NOT NULL DEFAULT '',
             added_by    INTEGER NOT NULL REFERENCES users(id),
             is_active   INTEGER NOT NULL DEFAULT 1,
@@ -90,7 +90,7 @@ async def create_tables(conn) -> None:
         CREATE TABLE IF NOT EXISTS ai_usage (
             id              INTEGER PRIMARY KEY AUTOINCREMENT,
             account_id      INTEGER NOT NULL REFERENCES accounts(id),
-            user_id         INTEGER NOT NULL DEFAULT 0,
+            user_id         BIGINT  NOT NULL DEFAULT 0,
             model           TEXT    NOT NULL DEFAULT '',
             request_type    TEXT    NOT NULL DEFAULT '',
             prompt_tokens   INTEGER NOT NULL DEFAULT 0,
@@ -112,7 +112,7 @@ async def create_tables(conn) -> None:
         CREATE TABLE IF NOT EXISTS ai_chat_history (
             id         INTEGER PRIMARY KEY AUTOINCREMENT,
             account_id INTEGER NOT NULL,
-            user_id    INTEGER NOT NULL,
+            user_id    BIGINT  NOT NULL,
             role       TEXT    NOT NULL,
             text       TEXT    NOT NULL,
             created_at TEXT    NOT NULL
@@ -132,7 +132,7 @@ async def create_tables(conn) -> None:
             visibility      TEXT    NOT NULL DEFAULT 'private',
             target_role     TEXT    NOT NULL DEFAULT 'all',
             pinned          INTEGER NOT NULL DEFAULT 0,
-            created_by      INTEGER NOT NULL DEFAULT 0,
+            created_by      BIGINT  NOT NULL DEFAULT 0,
             creator_name    TEXT    NOT NULL DEFAULT '',
             approved        INTEGER NOT NULL DEFAULT 1,
             updated_at      TEXT    NOT NULL DEFAULT '',
@@ -153,7 +153,7 @@ async def create_tables(conn) -> None:
             role            TEXT    NOT NULL,
             company_id      INTEGER,
             permissions     TEXT    NOT NULL DEFAULT '{}',
-            updated_by      INTEGER NOT NULL DEFAULT 0,
+            updated_by      BIGINT  NOT NULL DEFAULT 0,
             updated_at      TEXT    NOT NULL DEFAULT '',
             UNIQUE(account_id, role, company_id)
         );
@@ -178,7 +178,7 @@ async def create_tables(conn) -> None:
             account_id  INTEGER NOT NULL REFERENCES accounts(id),
             truck_num   TEXT    NOT NULL,
             is_primary  INTEGER NOT NULL DEFAULT 0,
-            assigned_by INTEGER NOT NULL DEFAULT 0,
+            assigned_by BIGINT  NOT NULL DEFAULT 0,
             assigned_at TEXT    NOT NULL DEFAULT ''
         );
         CREATE UNIQUE INDEX IF NOT EXISTS idx_driver_trucks_user_truck
@@ -191,7 +191,7 @@ async def create_tables(conn) -> None:
             user_id     INTEGER NOT NULL REFERENCES users(id),
             account_id  INTEGER NOT NULL REFERENCES accounts(id),
             company_id  INTEGER NOT NULL REFERENCES companies(id),
-            assigned_by INTEGER NOT NULL DEFAULT 0,
+            assigned_by BIGINT  NOT NULL DEFAULT 0,
             assigned_at TEXT    NOT NULL DEFAULT ''
         );
         CREATE UNIQUE INDEX IF NOT EXISTS idx_user_companies_user_company
